# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GenealogyPerson {
  id: string;
  /** @nullable */
  nepaliName?: string | null;
  englishName: string;
  role: string;
  initials: string;
  generation: number;
  branch: string;
  deceased: boolean;
  verificationStatus: string;
}

export interface GenealogyRelationship {
  id: string;
  personAId: string;
  personBId: string;
  type: string;
}

export interface GenealogyTree {
  centerPersonId: string;
  people: GenealogyPerson[];
  relationships: GenealogyRelationship[];
}

export interface RelationshipResult {
  connected: boolean;
  relationship: string;
  path: string[];
  hops: number;
  message: string;
}

export type GetGenealogyTreeParams = {
centerPersonId?: string;
/**
 * @minimum 1
 * @maximum 5
 */
generations?: number;
};

export type SearchPeopleParams = {
/**
 * @minLength 1
 */
q: string;
};

export type FindRelationshipParams = {
fromPersonId: string;
toPersonId: string;
};