# lib/api-zod/src/generated/types/genealogyPerson.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenealogyPerson {
  id: string;
  /** @nullable */
  nepaliName?: string | null;
  englishName: string;
  role: string;
  initials: string;
  generation: number;
  branch: string;
  deceased: boolean;
  verificationStatus: string;
}