# lib/api-zod/src/generated/types/genealogyTree.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenealogyPerson } from './genealogyPerson';
import type { GenealogyRelationship } from './genealogyRelationship';

export interface GenealogyTree {
  centerPersonId: string;
  people: GenealogyPerson[];
  relationships: GenealogyRelationship[];
}