# lib/api-zod/src/generated/types/relationshipResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RelationshipResult {
  connected: boolean;
  relationship: string;
  path: string[];
  hops: number;
  message: string;
}